script_python/spark/requirements.txt
pyspark
mysql-connector-python
pandas
pyarrow
fastparquet